(#5) Use filesystem backing for Brain.Karma
Prior to this commit, the karma brain would forget everything if the
process was stopped/restarted. This commit creates a backup file,
that is human-readable, on each karma change (this is terrible, but
will work for now). The process also attempts to read the karma
backup file on start up.

Close #5.